requirements.txt
# Core libraries
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.1
matplotlib==3.8.0
seaborn==0.12.2
tqdm==4.66.1
pyyaml==6.0

# PyTorch and Transformers
torch==2.2.0
torchvision==0.17.1
torchaudio==2.2.2
transformers==4.35.0

# Optional: for Jupyter notebooks
ipython==8.18.0
notebook==7.3.4
